Fix bugs

+++ edit_message.php
@@ -32,7 +32,11 @@
     if (isset($_POST['delete_message']))
     {
         $message = filter_input(INPUT_POST, 'message_id');
-        $db->query('DELETE FROM images WHERE id_message="'.$message.'"');
+        if($_SESSION['image'])
+        {
+            $db->query('DELETE FROM images WHERE id_message="'.$message.'"');
+            unset($_SESSION['image']);
+        }
         $db->query('DELETE FROM messages WHERE messages.id="'.$message.'"');
         header('Location: edit_message.php');
     }
@@ -59,6 +63,7 @@
             </form>
         </div>';
         $_SESSION['edit_string'] = true;
+        unset($_SESSION['image']);
     }
     if (isset($_POST['id_message']))
     {
@@ -75,6 +80,7 @@
             WHERE id='.$id_message);
             $message_query->bindValue(':message', $new_message, PDO::PARAM_STR);
             $message_query->execute();
+            unset($_SESSION['image']);
             header('Location: edit_message.php');
         }
         else if($_SESSION['user_power']<=2 && $rank<=1)
@@ -84,6 +90,7 @@
             WHERE id='.$id_message);
             $message_query->bindValue(':message', $new_message, PDO::PARAM_STR);
             $message_query->execute();
+            unset($_SESSION['image']);
             header('Location: edit_message.php');
         }
         else if ($_SESSION['user_power']<6 && $_SESSION['user_power']>2)
@@ -93,13 +100,16 @@
             WHERE id='.$id_message);
             $message_query->bindValue(':message', $new_message, PDO::PARAM_STR);
             $message_query->execute();
+            unset($_SESSION['image']);
             header('Location: edit_message.php');
         }
         else
         {
+            unset($_SESSION['image']);
             $_SESSION['message_error'] = '<div class="error">Brak uprawnień!</div>';
         }
     }
+    unset($_SESSION['image']);
 ?>
 <!DOCTYPE HTML>
 <html lang="pl">
@@ -112,11 +122,19 @@
         <meta name="author" content="Damian Zamroczynski" />
         <link rel="stylesheet" href="css/fontello.css">
         <link rel="stylesheet" href="css/bootstrap.min.css" />
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/baguettebox.js/1.8.1/baguetteBox.min.css">
         <link rel="stylesheet" href="css/main.css" />
         
         <link href="https://fonts.googleapis.com/css?family=Lato:400,700" rel="stylesheet">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
+        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" 
+                integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" 
+                crossorigin="anonymous"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" 
+                integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" 
+                crossorigin="anonymous"></script>
+        <script src="js/bootstrap.min.js"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/baguettebox.js/1.8.1/baguetteBox.min.js"></script>
         <!--[if lt IE 9]>
         <script src="//cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.min.js"</scripts>
         <![endif]-->
@@ -242,6 +260,7 @@
                                                         echo '</a>';
                                                         echo '</div>';
                                                         echo '</div>';
+                                                        $_SESSION['image'] = true;
                                                     }
                                                 }
                                                 echo '</div>';
@@ -300,12 +319,9 @@
         <footer>
             Stacja 4449 Bydgoszcz by Damian Zamroczynski &copy; 2019 Kontakt: damianzamroczynski@gmail.com
         </footer>
-        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" 
-                integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" 
-                crossorigin="anonymous"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" 
-                integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" 
-                crossorigin="anonymous"></script>
-        <script src="js/bootstrap.min.js"></script>
+        
+        <script>
+            baguetteBox.run('.tz-gallery');
+        </script>
     </body>
 </html>
